Deploy, deploy

+++ 2-do_deploy_web_static.py
@@ -12,24 +12,41 @@ def do_deploy(archive_path):
     if not path.isfile(archive_path):
         return False
 
-    try:
-        archive_name = archive_path[9:-4]
-        put(local_path=archive_path, remote_path='/tmp/{}.tgz'.format(
-            archive_name))
-        run('mkdir -p  /data/web_static/releases/{}'.format(archive_name))
-        run('tar -xzf /tmp/{}.tgz -C /data/web_static/releases/{}'.format(
-            archive_name, archive_name))
-        run('mv /data/web_static/releases/{}/web_static/* \
-             /data/web_static/releases/{}/'.format(archive_name, archive_name))
-        run('rm -rf /data/web_static/releases/{}/web_static'.format(
-            archive_name))
-        run('rm -rf /tmp/{}.tgz'.format(archive_name))
-        run('rm -rf /data/web_static/current')
-        run('ln -sf /data/web_static/releases/{} \
-            /data/web_static/current'.format(archive_name))
-    except Exception as e:
-        print("Deployment failed!!\n{}".format(e))
-        return False
+    archive_name = archive_path[9:-4]
+    
+   result = put(local_path=archive_path, remote_path='/tmp/{}.tgz'.format(
+        archive_name))
+   if result.failed:
+       return False
+
+    run('mkdir -p  /data/web_static/releases/{}'.format(archive_name))
+    result = run('tar -xzf /tmp/{}.tgz -C /data/web_static/releases/{}'.format(
+        archive_name, archive_name))
+   if result.failed:
+       return False
+
+    result = run('mv /data/web_static/releases/{}/web_static/* \
+         /data/web_static/releases/{}/'.format(archive_name, archive_name))
+   if result.failed:
+       return False
+
+    result = run('rm -rf /data/web_static/releases/{}/web_static'.format(
+        archive_name))
+   if result.failed:
+       return False
+
+    result = run('rm -rf /tmp/{}.tgz'.format(archive_name))
+   if result.failed:
+       return False
+
+    result = run('rm -rf /data/web_static/current')
+   if result.failed:
+       return False
+
+    result = run('ln -sf /data/web_static/releases/{} \
+        /data/web_static/current'.format(archive_name))
+   if result.failed:
+       return False
 
     print('New version deployed!')
     return True
